refactor sensei announcements
